% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Many sample means}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{a numeric vector, to be selected from}

\item{n}{the number of entries in each sample}

\item{reps}{integer, how many samples to take}
}
\value{

}
\description{
Many sample means
}
\examples{
many_sample_means(c(1,2,3,4,5), 4, 200)
}
