% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{sample_mean}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{a numeric vector, to be selected from}

\item{n}{the number of entries in the sample}
}
\value{

}
\description{
sample_mean
}
\examples{
sample_mean(c(1,2,3,4,5),5)

}
