% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Sample Means Ns}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{a numeric vector, to be selected from}

\item{reps}{integer, quantity of times to sample for each value of ns}

\item{ns}{vector of integers, sample sizes}
}
\value{

}
\description{
Sample Means Ns
}
